@@ -7,22 +7,44 @@ import {
 } from 'actions/user';
 
 export default function Main() {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
   const userInfo = useSelector((state: any) => state.user);
   const dispatch = useDispatch();
   const history = useHistory();
   useEffect(() => {
     dispatch(checkLoginAction());
   }, []);
+
+  const hanleInputUsername = e => {
+    setUsername(e.currentTarget.value);
+  };
+
+  const handleInputPassword = e => {
+    setPassword(e.currentTarget.value);
+  };
+
   return (
     <section className="">
       <div>{userInfo && userInfo.username}</div>
 
+      <input
+        placeholder="请输入用户名"
+        value={username}
+        onChange={hanleInputUsername}
+      />
+      <input
+        placeholder="请输入密码"
+        value={password}
+        onChange={handleInputPassword}
+      />
       <button
         onClick={() => {
-          dispatch(loginAction('wyc', '123'));
+          dispatch(loginAction(username, password));
         }}>
         登录
       </button>
+      <div></div>
       <button
         onClick={() => {
           history.push('/');
